only owner can remove comment

+++ Infrastructure/GhostNetwork.Infrastructure.Repository/NewsFeedStorage.cs
@@ -177,9 +177,15 @@ namespace GhostNetwork.Infrastructure.Repository
             return comment == null ? null : ToDomain(comment);
         }
 
-        public async Task DeleteCommentAsync(string id)
+        public async Task<bool> DeleteCommentAsync(string id)
         {
-            await commentsApi.DeleteAsync(id);
+            var comment = await commentsApi.GetByIdAsync(id);
+            if (comment.Author.Id.ToString() == currentUserProvider.UserId)
+            {
+                await commentsApi.DeleteAsync(id);
+                return true;
+            }
+            return false;
         }
 
         private static long GetTotalCountHeader<T>(Publications.Client.ApiResponse<T> response)
